+++ eos-updater-flatpak-installer/main.c
@@ -23,7 +23,7 @@
 #include <flatpak.h>
 #include <glib.h>
 #include <libeos-updater-util/enums.h>
-#include <libeos-updater-util/flatpak.h>
+#include <libeos-updater-util/flatpak-util.h>
 #include <libeos-updater-util/types.h>
 #include <libeos-updater-util/util.h>
 #include <libeos-updater-flatpak-installer/installer.h>
