+++ eos-updater/fetch.c
@@ -24,7 +24,7 @@
 #include <eos-updater/fetch.h>
 #include <eos-updater/object.h>
 #include <flatpak.h>
-#include <libeos-updater-util/flatpak.h>
+#include <libeos-updater-util/flatpak-util.h>
 #include <libeos-updater-util/types.h>
 #include <libeos-updater-util/util.h>
 #include <libmogwai-schedule-client/schedule-entry.h>
