+++ libeos-updater-flatpak-installer/perform-flatpak-actions.c
@@ -24,7 +24,7 @@
 #include <glib.h>
 #include <libeos-updater-flatpak-installer/installer.h>
 #include <libeos-updater-util/enums.h>
-#include <libeos-updater-util/flatpak.h>
+#include <libeos-updater-util/flatpak-util.h>
 #include <libeos-updater-util/types.h>
 #include <libeos-updater-util/util.h>
 #include <locale.h>
