+++ libeos-updater-util/flatpak-util.h
