+++ libeos-updater-util/ostree-util.h
